+++ CHANGES.md
@@ -2,6 +2,8 @@
 
 ### Release v4.4
 * Move legacy style config settings to plugin config settings
+* Remove toggle button for debug settings
+* Show an infobox if the user debug mode is active
 
 ### Release v4.3.1
 * Check the realuser from \core\session\manager::get_realuser() to have debugging in "loginas" sessions too.